Fixed this bug:
  valgrind: vg_cachesim.c:389 (get_BBCC): Assertion `((Bool)0) == remove' failed.


Problem was that when an exe segment was unloaded, cachesim_notify_munmap() was
being called after symbols were unloaded.  But it needs the symbols to do the
lookup required to remove the BBCCs.  It was only working some of the time
for exe segments that didn't have any symbols(!)

Fix:  now invalidate translations first, unload symbols second.  This required
adding VG_(is_munmap_exe)() to determine if an unloaded segment is executable.


git-svn-id: svn://svn.valgrind.org/valgrind/branches/unlabeled-1.76.2@839 a5019735-40e9-0310-863c-91ae7b9d1cf9