Introduced an 'extra' field for malloc'd blocks.  Can be of arbitary size.
Currently only used to store the requested size in bytes for blocks in arena
VG_AR_CLIENT, as a preliminary step to allow me to move all the ShadowChunk
stuff out of core into the memcheck skin.

Also added functions VG_({get,set}_clientblock_req_pszB) for accessing the
field, which is used in vg_clientmalloc.c.  Tested it carefully to make sure
that returned req_pszB always matches the size in the corresponding
ShadowChunk.


git-svn-id: svn://svn.valgrind.org/valgrind/branches/unlabeled-1.6.6@916 a5019735-40e9-0310-863c-91ae7b9d1cf9