Fixed up the messiness around blocking syscalls a bit.
Split up VG_(check_known_blocking_syscall) into two functions, one each for the
pre- and post-actions.  This unfortunately logically separates the pre- and
post-actions in the source, but the alternative would have required absurd
arm-twisting, especially with the returning of the pre_result back to
vg_scheduler.c.

Added a 'pre_result' field to the VgWaitedOnFd table to store the pre_result
from pre-checking of blocking syscalls.


git-svn-id: svn://svn.valgrind.org/valgrind/branches/unlabeled-1.76.4@926 a5019735-40e9-0310-863c-91ae7b9d1cf9