Merge revs:    memcheck/docs mc_main.html,1.7,1.8    helgrind helgrind.h,1.4,1.5    include valgrind.h,1.18,1.19    memcheck memcheck.h,1.12,1.13    coregrind/docs coregrind_core.html,1.10,1.11
Fixed up the client request info in the docs, which required splitting
the core requests from the Memcheck requests and putting the
descriptions in the appropriate parts of the docs.

Removed the __VALGRIND_SOME_SKIN_H test -- it was designed to prevent
people #including valgrind.h when they probably want memcheck.h, but
it's actually a reasonable thing to do to #include valgrind.h alone,
because there are some requests in valgrind.h.

Removed references to VALGRIND_MAKE_NOACCESS_STACK from the docs, as
it doesn't exist in the implementation.

Fixed a few other minor things.


git-svn-id: svn://svn.valgrind.org/valgrind/branches/VALGRIND_2_0_REALLY@1890 a5019735-40e9-0310-863c-91ae7b9d1cf9