Merge revs:   addrcheck  ac_main.c       1.52   cachegrind cg_main.c       1.53              cg_sim_gen.c    1.10   memcheck   mc_main.c       1.39   helgrind   hg_main.c       1.62   coregrind  vg_from_ucode.c 1.64              vg_libpthread.c 1.134              vg_symtab2.c    1.51              vg_translate.c  1.58
Delete some "inline" directives on functions that gcc-3.3.1 declined
to inline.  This is needed to get a warning-free compilation on 3.3.1.
It seems we had "inline" on some pretty huge functions in places.
Also it appears gcc-3.3.1 won't inline a function call in a tail call
position, reasonably enough.  I assume in that case it prefers to
create a tailcall to the callee, rather than inlining it.


git-svn-id: svn://svn.valgrind.org/valgrind/branches/unlabeled-1.130.2@1911 a5019735-40e9-0310-863c-91ae7b9d1cf9