Nasty build-system hackery to get coregrind/ built as a library against which tools are statically linked.  Currently what is here at the moment is able to construct a statically linked (?) binary valgrind_memcheck and install it in $(prefix)/bin.  Note, the resulting binary does not work yet.
There is an appalling kludge which would benefit from the attention of
people more skilled than me in the Art Of Automake:

In coregrind, I fundamentally want automake to build 'libcoregrind.a'.
And automake is all set up to do that.  Except .. automake constructs
the .a by using /usr/bin/ar, which is a problem because some of the
inputs to libcoregrind.a are themselves .a files (specifically, the
results of building any m_modulename/ directory).  And so ar barfs.

My horrible hack is tell automake that the name of 'ar' is in fact 'ld
-r -o'.  Since ld is able to accept both .o and .a files as input this
works, except that the resulting libcoregrind.a is actually a .o file.

If anybody knows how to fix this properly I would be happy to hear
from you.

Other changes:

- renamed all aspects of the magic linker script 'stage2.lds' to
  'valt_load_address.lds' and lift it up to the top level directory,
  since all tools will need to see it when they are linked.

- got rid of everything to do with PIE

Other kludges:

- load_tool() in m_main.c is now kludged to (1) just return &VG_(tool_info)
  and (2) to return a hardwired preloadpath.  These will need to be
  fixed properly.





git-svn-id: svn://svn.valgrind.org/valgrind/branches/ASPACEM@4342 a5019735-40e9-0310-863c-91ae7b9d1cf9