Commit initial framework changes aimed at making data symbol reading and .got/.plt identification work properly on Linux.
The basic observation is that until now, Valgrind has waited until an
ELF file is mapped r-x, then attempted to read both text and data
symbols for the file.  That's reasonable for the text symbols but is
nonsense for the data symbols, since we have no idea where ld.so will
later map the file's data section.

The basic change is to cause debuginfo.c to wait to see both a r-x and
a rw- mapping for an ELF object file, before reading its symbols.
Only that way will we know where the data segment is mapped to, and
hence can we know what the data symbol actual VMAs are.

As well as revising the when-to-load-symbols logic, much accumulated
cruft in readelf.c is being tidied up.

Note this is work in progress and is majorly broken at the moment.




git-svn-id: svn://svn.valgrind.org/valgrind/branches/DATASYMS@7319 a5019735-40e9-0310-863c-91ae7b9d1cf9