Add a parser for DWARF3 .debug_info, .debug_loc and .debug_range sections, intended to extract variable type and location information and format the information in such a way as to support rapid querying later.
The existing DWARF3 "Other Debugging Information" reader
(readdwarf.c), which extracts line number and frame unwind
information, is unaffected (so as not to break it).  In time
perhaps the two readers can be integrated.

This code parses large (eg 80MB) files fairly quickly.  Its speed
could probably be doubled with a little further effort.

It is minimally tested and fairly untidy.  It will be refactored
shortly.



git-svn-id: svn://svn.valgrind.org/valgrind/branches/DATASYMS@7380 a5019735-40e9-0310-863c-91ae7b9d1cf9