Generate instrumentation code to shadow most values with 32-bit origin tags (referred to by the generic letter 'b' in the code, as opposed to 'v' for the definedness shadow values).
Improve debug tracing, so as to make it easy to see which code is for
V-bit (definedness) tracking and which for B-tag (origin) tracking.

The important functions are schemeS, which generates origin tracking
code for an IR statement, and schemeE, which generates an origin
expression for an IR expression.

The existing V-bits tracking code is almost completely unchanged.
The only change is that when generating a call to
MC_(helperc_value_check0_fail), schemeE is used to generate an origin
value to pass to the helper function.  (Obviously, otherwise all of
this would be completely pointless).

This involves some nasty x86-specific hacks to do with guest state
layout, which will have to be cleaned up for real.  See function
get_shadow_offset for the gruesome details.



git-svn-id: svn://svn.valgrind.org/valgrind/branches/OTRACK_BY_INSTRUMENTATION@7811 a5019735-40e9-0310-863c-91ae7b9d1cf9