Further abstractify the concept of 'origin tag' as seen by the core-tool interface, so that the interface only 'knows' about ECUs: ExeContext Uniques.  These are 30-bit numbers which have a 1-1 mapping with (valid) ExeContext*s.
The 'origin tag' concept becomes purely Memcheck internal.  Memcheck
now creates origin tags by taking an ECU and putting its own
information in the lowest 2 bits.

The point of all this is to be able to encode, in each origin tag, a
2-bit value which indicates whether the origin is from a stack
allocation, a heap allocation, a client request
(VALGRIND_MAKE_MEM_UNDEFINED) or from some other kind of source, for
example sys_brk.  This means it can classify the origin reliably in
error messages presented to the user.

As a result, mc_main.c's kludges re "is_stack_origin" have been
removed.

Also as a result, the number of ExeContexts representable falls to
2^30 (1 billion).  That does not sound like a big limitation in
practice: on a 32-bit machine it's a non-problem.  On a 64-bit
platform, each ExeContext will take a minimum of 24 bytes + malloc
overhead, which means representing 1 billion ExeContexts will soak up
40GB of memory.



git-svn-id: svn://svn.valgrind.org/valgrind/branches/OTRACK_BY_INSTRUMENTATION@7956 a5019735-40e9-0310-863c-91ae7b9d1cf9