A small addition to the core/tool interface, needed to support the SGcheck tool.  If affects only the "new_mem_startup" and "new_mem_mmap" events.  When a new mmap happens, and for the equivalent startup events, the debuginfo reader may choose to load debug info associated with the mapped object.
This commit adds an extra parameter to the abovementioned events,
which is a 64-bit abstract handle that can later be used to refer
to the associated debuginfo that got loaded.  This allows tools
to be aware of when debuginfo gets loaded, so that they can make
further queries to m_debuginfo if they like, before allowing execution
of the client to continue.

Patch is not cleanly separated from other changes (unfortunately)
and so is not standalone.



git-svn-id: svn://svn.valgrind.org/valgrind/branches/SGCHECK@8529 a5019735-40e9-0310-863c-91ae7b9d1cf9